+++ src/imio/smartweb/core/viewlets/ogptags.py
@@ -31,7 +31,7 @@ class OgpTagsViewlet(HeaderViewlet):
 
     def set_ogp_informations_for_view(self):
         self._item = {
-            "title": self.context.Title,
+            "title": self.context.Title(),
             "description": self.context.description,
             "url": self.context.absolute_url(),
         }
@@ -68,24 +68,15 @@ class OgpTagsViewlet(HeaderViewlet):
         if IDirectoryView.providedBy(self.context):
             params = "fullobjects=1"
             auth_source_url = DIRECTORY_URL
-            client_id = os.environ.get("RESTAPI_DIRECTORY_CLIENT_ID")
-            client_secret = os.environ.get("RESTAPI_DIRECTORY_CLIENT_SECRET")
         elif IEventsView.providedBy(self.context):
             params = "metadata_fields=category&metadata_fields=local_category&metadata_fields=container_uid&metadata_fields=topics&metadata_fields=start&metadata_fields=end&metadata_fields=has_leadimage&metadata_fields=UID&fullobjects=1"
             auth_source_url = EVENTS_URL
             endpoint = "@events"
-            client_id = os.environ.get("RESTAPI_EVENTS_CLIENT_ID")
-            client_secret = os.environ.get("RESTAPI_EVENTS_CLIENT_SECRET")
         elif INewsView.providedBy(self.context):
             params = "fullobjects=1"
             auth_source_url = NEWS_URL
-            client_id = os.environ.get("RESTAPI_NEWS_CLIENT_ID")
-            client_secret = os.environ.get("RESTAPI_NEWS_CLIENT_SECRET")
-        auth = get_wca_token(client_id, client_secret)
-        if not auth:
-            return
         auth_source_url = f"{auth_source_url}/{endpoint}?UID={uid}&{params}"
-        result_json = get_json(auth_source_url, auth=auth)
+        result_json = get_json(auth_source_url)
         if result_json:
             self._item = result_json["items"][0]
             self._set_image()